identity-service/pyproject.toml
[project]
name = "identity-service"
version = "0.1.0"
description = "Identity Service for Potion. Handles user authentication and management."
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "alembic>=1.17.2",
    "argon2-cffi>=25.1.0",
    "fastapi[standard]>=0.122.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.44",
]
